重构了仓库模块，为供货商客户类型的删除操作增加关联检查，同时将部分窗体打开方式更改为对话框形式

+++ TypeControl/ClientTypeForm.cs
@@ -50,11 +50,19 @@ namespace 仓库管理系统
         {
             List<DataRow> dataRows = MDIAction.GetGridViewCheckedRows(dataGridView1);
             List<TClientType> clientTypes = TypeControlAction.DataRowToClientType(dataRows);
-            int rows = TypeControlQuery.DeleteClientTypeInfo(clientTypes);
-            MessageBox.Show($"成功删除{rows}行");
-            TypeControlAction.HideLabTxtBtn(typeNameLab, typeNameTxt, typeRankLab, typeRankTxt, insertBtn, alterBtn);
-            saveBtn.Visible = true;
-            FlashForm();
+            if (MDIAction.CheckIsAllowDelClientType(clientTypes))
+            {
+                int rows = TypeControlQuery.DeleteClientTypeInfo(clientTypes);
+                MessageBox.Show($"成功删除{rows}行");
+                TypeControlAction.HideLabTxtBtn(typeNameLab, typeNameTxt, typeRankLab, typeRankTxt, insertBtn, alterBtn);
+                saveBtn.Visible = true;
+                FlashForm();
+            }
+            else
+            {
+                MessageBox.Show("选中类型仍有关联不可删除");
+            }
+
         }
 
         public override void insertBtn_Click(object sender, EventArgs e)
